Solution to Logical AND Operation

+++ Logical AND Operation@@67882a731b4392d8fc1577dc/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    double a,b;
-    scanf("%lf %lf",&a,&b);
+    int a,b;
+    scanf("%d %d",&a,&b);
     if(a>0 && b>0){
         printf("True");
     }
